+++ source/skins/simple/FormFieldSkin.js
@@ -5,7 +5,7 @@ import classnames from 'classnames';
 import type { FormFieldProps } from '../../components/FormField';
 import { PopOver } from '../../components/PopOver';
 import { SimpleFormFieldVariables } from '../../themes/simple/SimpleFormField';
-import { handleRefFocusState } from '../../utils/hooks';
+import { handleRefState, manageRef } from '../../utils/hooks';
 
 type Props = FormFieldProps & {
   formFieldRef: ElementRef<*>,
@@ -14,11 +14,15 @@ type Props = FormFieldProps & {
 };
 
 export function FormFieldSkin(props: Props) {
-  const [isFormFieldFocused, setIsFormFieldFocused] = useState(false);
-  const setFormFieldRef = handleRefFocusState(
-    props.formFieldRef,
-    setIsFormFieldFocused
-  );
+  const updateOnRefChanges = manageRef(props.formFieldRef);
+  const isFormFieldFocused = handleRefState(props.formFieldRef, {
+    on: 'focus',
+    off: 'blur',
+  });
+  const isFormFieldHovered = handleRefState(props.formFieldRef, {
+    on: 'mouseenter',
+    off: 'mouseleave',
+  });
   const hasError = props.error != null;
   return (
     <div
@@ -28,6 +32,7 @@ export function FormFieldSkin(props: Props) {
         props.disabled ? props.theme[props.themeId].disabled : null,
         props.error ? props.theme[props.themeId].errored : null,
       ])}
+      style={props.themeVariables}
     >
       {props.label && (
         <label
@@ -41,19 +46,23 @@ export function FormFieldSkin(props: Props) {
       )}
       <PopOver
         visible={
-          props.isErrorShown ||
-          (hasError && isFormFieldFocused && !props.isErrorHidden)
+          props.isErrorShown === true ||
+          (props.isErrorHidden !== true &&
+            hasError &&
+            ((props.isShowingErrorOnFocus && isFormFieldFocused) ||
+              (props.isShowingErrorOnHover && isFormFieldHovered)))
         }
         content={props.error}
         themeVariables={{
           '--rp-pop-over-bg-color': `var(${SimpleFormFieldVariables.errorColor}`,
+          ...props.themeVariables,
         }}
         placement="bottom"
         popperOptions={{ modifiers: [{ name: 'flip', enabled: false }] }}
         duration={[300, 0]}
       >
         <div className={props.theme[props.themeId].inputWrapper}>
-          {props.render(setFormFieldRef)}
+          {props.render(updateOnRefChanges)}
         </div>
       </PopOver>
     </div>